pluginsdk: move TargetRef util to exported pkg
 Description
Moves the TargetRefs conversion utils to an exported
pkg to allow external plugins to use it.

 Change Type

```
/kind cleanup
```

 Changelog

```release-note
NONE
```
