+++ pkg/pluginsdk/utils/policy.go
@@ -1,4 +1,4 @@
-package pluginutils
+package utils
 
 import (
 	"maps"
